Adding support for labels.
(Forgive me as im not the best coder lol)

But I wanted to add the ablility to label hex codes or icao numbers in the "watchlist.txt"
and have the labels be sent into telegram.

the structure for the watchlist.txt stays mostly the same but now it should look like this: "a88bde: Dept of Commerce (hurricane hunters)"

all that needs to be added to send a label is adding ": " at the end of an icao or hex.

this is pretty crude but gets the job done! but it way need some tweaking (again pretty new to coding).